transaction successfully saving to database

+++ db/spending_tracker.sql
@@ -19,5 +19,5 @@ CREATE TABLE transactions
     id SERIAL primary key,
     merchant_id INT REFERENCES merchants(id),
     category_id INT REFERENCES categories(id),
-    amount INT
+    amount decimal(18,2)
 );
